+++ plugins/spectral_compressor/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Reworked the resetting behavior again to smoothly fade the timings in after a
+  reset to allow the envelop followers to more gently settle in.
+
 ## [0.4.1] - 2023-03-22
 
 ### Fixed